refactor: Stub -> Fake 네이밍 수정
